feat: :zap: adding user logic to the game (adding login and signin)

+++ apps/backend/src/lobby/lobby.controller.ts
@@ -1,6 +1,7 @@
 import {
   Body,
   Controller,
+  Delete,
   Get,
   HttpCode,
   NotFoundException,
@@ -11,13 +12,42 @@ import {
 import { LobbyService } from './lobby.service';
 import { Lobby, Prisma, Player, Dificulty } from '../generated/prisma/client';
 import { PlayerService } from '../player/player.service';
+import { UserService } from '../user/user.service';
+
+const PRESET_COLORS = ["#ef4444", "#3b82f6", "#eab308", "#22c55e"];
 
 @Controller('lobby')
 export class LobbyController {
   constructor(
     private readonly lobbyService: LobbyService,
     private readonly playerService: PlayerService,
+    private readonly userService: UserService
   ) {}
+
+
+  @Put('join')
+  async joinLobby(
+    @Body() joinData: { code: string, userId: number }
+  ): Promise<Lobby> {
+    const lobby = await this.lobbyService.getLobbyByCode(joinData.code);
+    if(lobby.numPlayers >= 4) {
+      throw new NotFoundException(`El lobby con código ${joinData.code} está lleno`);
+    }
+    const user = await this.userService.getUser(Number(joinData.userId));
+    const data = {
+      name: user.username,
+      color: PRESET_COLORS[lobby.numPlayers],
+      movement: 3,
+      turnOrder: lobby ? lobby.numPlayers : 0,
+      lobby: { connect: { id: Number(lobby.id) } },
+      user: { connect: { id: Number(joinData.userId) } },
+    };
+    const newPlayer = await this.playerService.createPlayer(data);
+    return this.lobbyService.addPlayerToLobby({
+      where: { id: Number(lobby.id) },
+      data: { playerId: newPlayer.id },
+    });
+  }
 
   @Get('/:id')
   async getLobby(@Param('id') id: string): Promise<Lobby> {
@@ -63,14 +93,17 @@ export class LobbyController {
   @Put('/:id/add-player')
   async addPlayerToLobby(
     @Param('id') id: string,
-    @Body() playerData: { name: string; color: string },
+    @Body('userId') userId: string ,
   ): Promise<Lobby> {
-    const lobby = await this.getLobby(id);
+    const lobby = await this.lobbyService.getLobby({ id: Number(id) });
+    const user = await this.userService.getUser(Number(userId));
     const data = {
-      name: playerData.name,
-      color: playerData.color,
+      name: user.username,
+      color: PRESET_COLORS[lobby.numPlayers],
       movement: 3,
       turnOrder: lobby ? lobby.numPlayers : 0,
+      lobby: { connect: { id: Number(id) } },
+      user: { connect: { id: Number(userId) } },
     };
     const newPlayer = await this.playerService.createPlayer(data);
     return this.lobbyService.addPlayerToLobby({
@@ -80,12 +113,18 @@ export class LobbyController {
   }
 
   @Put('/:id/remove-player')
-  async removePlayerFromLobby(@Param('id') playerId: string): Promise<Lobby> {
+  async removePlayerFromLobby(@Param('id') playerId: string, @Body() userId: number): Promise<Lobby> {
     try {
       const lobby = await this.lobbyService.getLobbieFromPlayer(
-        Number(playerId),
+        Number(playerId)
       );
 
+      const player = await this.playerService.getPlayer({ id: Number(playerId) });
+      if(lobby.hostId !== Number(userId) && player.userId !== Number(userId)) {
+        throw new NotFoundException(
+          `El jugador no puede ser eliminado`,
+        );
+      }
       await this.playerService.deletePlayer({ id: Number(playerId) });
       if (!lobby) {
         throw new NotFoundException(
@@ -102,4 +141,16 @@ export class LobbyController {
       );
     }
   }
+
+  @Delete('/:id')
+  async deleteLobby(@Param('id') id: string, @Body() userId: Number): Promise<Lobby> {
+    const lobby = await this.lobbyService.getLobby({ id: Number(id) });
+    if(lobby.hostId !== Number(userId)) {
+      throw new NotFoundException(
+        `El usuario no tiene permisos para eliminar el lobby ${Number(id)}`,
+      );
+    }
+    return this.lobbyService.deleteLobby(Number(id));
+  }
+
 }
